added useBackdrop
