+++ src/components/widgets/Backdrop/Backdrop.module.css
@@ -3,15 +3,6 @@
     width: 100%;
 }
 
-.backdropBG {
-    position: absolute;
-    left: -260px;
-    width: "99.3vw";
-    background-color: white;
-    z-index: 1;
-    transition: "all 0.3s ease";
-}
-
 .containerRef {
     position: relative;
     z-index: 2;
